SO-2185: moving RF2 export client implementation to rf2.internal package
