+++ snomed/com.b2international.snowowl.snomed.api.impl/META-INF/MANIFEST.MF
@@ -19,7 +19,6 @@ Require-Bundle: org.eclipse.core.runtime,
  com.b2international.snowowl.snomed.api,
  com.b2international.snowowl.snomed.common,
  com.b2international.snowowl.snomed.exporter.server,
- com.b2international.snowowl.snomed.exporter,
  com.b2international.snowowl.snomed.dsl,
  com.b2international.snowowl.snomed.importer,
  com.b2international.snowowl.snomed.importer.rf2,
