+++ snomed/com.b2international.snowowl.snomed.datastore/META-INF/MANIFEST.MF
@@ -55,6 +55,7 @@ Export-Package: com.b2international.snowowl.snomed.core.domain,
  com.b2international.snowowl.snomed.datastore.index.refset,
  com.b2international.snowowl.snomed.datastore.index.update,
  com.b2international.snowowl.snomed.datastore.internal.id.reservations,
+ com.b2international.snowowl.snomed.datastore.internal.rf2,
  com.b2international.snowowl.snomed.datastore.model,
  com.b2international.snowowl.snomed.datastore.request,
  com.b2international.snowowl.snomed.datastore.request.rf2,
