+++ snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/internal/rf2/AbstractSnomedDsvExportItem.java
@@ -13,13 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.snomed.exporter.model;
+package com.b2international.snowowl.snomed.datastore.internal.rf2;
 
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
-
-import com.b2international.snowowl.snomed.exporter.service.SnomedRefSetDSVExportClientRequest;
 
 /**
  * Represents an exported column in DSV files created from reference sets of the SNOMED&nbsp;CT terminology.
