+++ snomed/com.b2international.snowowl.snomed.exporter.server/META-INF/MANIFEST.MF
@@ -9,7 +9,6 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.9.0",
  org.eclipse.net4j;bundle-version="[4.1.1,4.1.2)",
  com.b2international.snowowl.snomed.datastore,
  com.b2international.snowowl.datastore.server,
- com.b2international.snowowl.snomed.exporter,
  com.b2international.snowowl.dependencies
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
