--- snomed/com.b2international.snowowl.snomed.exporter/META-INF/MANIFEST.MF
@@ -1,15 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Snow Owl SNOMED CT Exporter Bundle
-Bundle-SymbolicName: com.b2international.snowowl.snomed.exporter
-Bundle-Version: 5.7.0.qualifier
-Bundle-Activator: com.b2international.snowowl.snomed.exporter.SnomedExporterActivator
-Bundle-Vendor: B2i Healthcare
-Require-Bundle: org.eclipse.core.runtime;bundle-version="3.9.0",
- org.eclipse.net4j;bundle-version="[4.1.1,4.1.2)",
- com.b2international.snowowl.snomed.datastore
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-ActivationPolicy: lazy
-Export-Package: com.b2international.snowowl.snomed.exporter.model,com.
- b2international.snowowl.snomed.exporter.service
-Import-Package: org.slf4j;version="1.7.2"
